prefinal test

+++ calculateur.sh
@@ -1,7 +1,55 @@
 #!/bin/bash
 
-apps=("firefox", "gummi", "anjuta", "libreoffice")
+#updating apt && installing software-properties-common
+sudo apt update
+sudo apt install software-properties-common
 
-for app in "${apps[@]}"; do
-	sudo apt install $app
+# Declare a list of software to propose for installation
+software_list=(
+  "firefox"
+  "vim"
+  "git"
+  "python3"
+  "netcat"
+)
+
+# Print a list of the proposed software
+echo "The following software is available for installation:"
+for app in "${software_list[@]}"; do
+  echo "  $app"
 done
+
+# Ask the user if they want to install all of the software
+read -p "Do you want to install all of the software? (y/n) " response
+
+# If the user says yes, install all of the software
+if [[ $response == "y" ]]; then
+  for app in "${software_list[@]}"; do
+    sudo apt install $app
+  done
+else
+  # If the user says no, let them select the software they want to install
+  while true; do
+    # Print a list of the software that has not yet been installed
+    echo "The following software has not yet been installed:"
+    for app in "${software_list[@]}"; do
+      if ! dpkg -l | grep -q "^ii $app"; then
+        echo "  $app"
+      fi
+    done
+
+    # Ask the user to select the software they want to install
+    read -p "Select the software you want to install (enter a space-separated list): " software
+
+    # Install the selected software
+    for app in $software; do
+      sudo apt install $app
+    done
+
+    # Check if the user wants to install more software
+    read -p "Do you want to install more software? (y/n) " response
+    if [[ $response != "y" ]]; then
+      break
+    fi
+  done
+fi
